+++ src/components/map/CampusMinimap.tsx
@@ -1,18 +1,20 @@
 import { useApp } from '@/app/Context';
 import { getCourseColor } from '@/app/Plan';
-import { getLocation } from '@/app/Schedule';
+import { getLocation, isRoomFinderAvailable } from '@/app/Schedule';
 import { Color } from '@/types/BaseTypes';
 import { ScheduleSection } from '@/types/ScheduleTypes';
 import { Mode } from '@/utility/Constants';
 import { ArrowsPointingOutIcon } from '@heroicons/react/24/outline';
 import { Map } from 'leaflet';
-import { Fragment, useEffect, useRef } from 'react';
-import { MapContainer, Marker, TileLayer, Tooltip } from 'react-leaflet';
+import { Fragment, useEffect, useMemo, useRef } from 'react';
+import { MapContainer, Marker, TileLayer } from 'react-leaflet';
+import MapTooltip from './MapTooltip';
 import {
   MapFlyTo,
   getMapMarkerIcon,
   getUnknownMapMarkerIcon,
 } from './MapUtility';
+import { MapIcon } from '@heroicons/react/24/solid';
 
 const DEFAULT_POSITION: [number, number] = [42.055909, -87.672709];
 const DEFAULT_ZOOM = 14.2;
@@ -24,7 +26,15 @@ interface CampusMinimapProps {
 
 function CampusMinimap({ location, section }: CampusMinimapProps) {
   const { userOptions, mapView } = useApp();
-  const rooms = location ? [location[0]] : section?.room || [];
+  const rooms = useMemo(
+    () => (location ? [location[0]] : section?.room || []),
+    [location, section?.room]
+  );
+  const roomFinderAvailable = useMemo(
+    () => rooms?.some((r) => isRoomFinderAvailable(r)),
+    [rooms]
+  );
+
   const ref = useRef<Map | null>(null);
 
   const positions: ([number, number] | null)[] | undefined = rooms.map(
@@ -46,6 +56,7 @@ function CampusMinimap({ location, section }: CampusMinimapProps) {
   }, [userOptions.get.mode, userOptions.get.tab]);
 
   const firstPosition = positions?.[0] || DEFAULT_POSITION;
+
   return (
     <div className="relative h-full w-full">
       <MapContainer
@@ -76,7 +87,14 @@ function CampusMinimap({ location, section }: CampusMinimapProps) {
                         )
                       : getMapMarkerIcon(location![1])
                   }
-                />
+                >
+                  <MapTooltip direction="bottom" permanent>
+                    <div className="flex items-center gap-1">
+                      <span>{rooms[i]}</span>
+                      {roomFinderAvailable && <MapIcon className="h-3 w-3" />}
+                    </div>
+                  </MapTooltip>
+                </Marker>
               ) : (
                 <Fragment key={`map-marker-${i}`} />
               )
@@ -86,9 +104,9 @@ function CampusMinimap({ location, section }: CampusMinimapProps) {
               position={DEFAULT_POSITION}
               icon={getUnknownMapMarkerIcon('gray')}
             >
-              <Tooltip permanent direction="bottom">
+              <MapTooltip direction="bottom" permanent>
                 Unknown location
-              </Tooltip>
+              </MapTooltip>
             </Marker>
           ))}
       </MapContainer>